requirements.txt
altair==4.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==22.2.0
backports.zoneinfo==0.2.1
backtrader==1.9.78.123
backtrader-plotting==2.0.0
baostock==0.8.8
beautifulsoup4==4.12.2
blinker==1.6.3
bokeh==2.3.3
bs4==0.0.1
cachetools==5.2.1
certifi==2023.5.7
charset-normalizer==3.2.0
chinese-calendar==1.9.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
comtypes==1.2.0
contourpy==1.1.0
cssselect==1.2.0
cycler==0.11.0
decorator==5.1.1
Django==3.1.12
django-filter==2.4.0
djangorestframework==3.12.4
easyutils==0.1.7
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.104.0
Flask==3.0.0
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
jqktrader==0.1.4
jsonschema==4.17.3
kiwisolver==1.4.5
lxml==4.9.3
markdown2==2.4.10
MarkupSafe==2.1.1
matplotlib==3.7.2
mysql-connector-python==8.1.0
mysqlclient==2.2.0
numpy==1.22.3
openpyxl==3.1.2
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
protobuf==4.21.12
pyarrow==10.0.1
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pypiwin32==223
pyquery==2.0.0
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinauto==0.6.8
PyYAML==6.0.1
requests==2.31.0
rich==13.0.1
semver==2.13.0
Shapely==1.8.4
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
starlette==0.27.0
streamlit==1.16.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tushare==1.2.89
typing_extensions==4.8.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.23.2
validators==0.20.0
watchdog==3.0.0
websocket-client==0.57.0
Werkzeug==3.0.0
zipp==3.11.0
